Refactor authentication services and update namespaces
- Updated `AuthStartupTask.cs` to use new extensions and handle configuration directly.
- Modified `CheckCreateUserAsync` to retrieve admin password via `MicroservicesExtensions`.
- Enhanced `IdentityService.cs` with asynchronous JWT options retrieval and token creation.
- Updated user permissions and modules handling in `LoginAsync`.
- Changed `RefreshTokenAsync` and `CreateToken` to asynchronous methods.
- Moved `IIdentityService` to the `Interfaces` namespace for better structure.
